Refactor: AgentSignalRepository 파라미터 타입 수정

+++ capstone1/src/main/java/org/vox/capstonedesign1/repository/AgentSignalRepository.java
@@ -9,7 +9,7 @@ public interface AgentSignalRepository extends JpaRepository<AgentSignal, Long> 
 //  CrudRepository에 구현되어있음
 //  AgentSignal save(AgentSignal signal);
 //  Optional<AgentSignal> findById(Long id);
-    List<AgentSignal> findAllByDevice_Id(int deviceId);
+    List<AgentSignal> findAllByDevice_Id(Long deviceId);
     // 데이터를 최신시간순으로 정렬해서 가져오기
     List<AgentSignal> findAllByOrderByTimeStampDesc();
 }
